requirements.txt
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
bcrypt==4.3.0
beautifulsoup4==4.13.3
bigquery==0.0.45
blinker==1.9.0
bs4==0.0.2
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudevents==1.11.0
cmdstanpy==1.2.5
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.1
cycler==0.12.1
dacktool==0.0.7
db-dtypes==1.4.1
dbstream==0.1.28
deprecation==2.1.0
ecdsa==0.19.0
et_xmlfile==2.0.0
fastapi==0.115.11
Flask==3.1.0
fonttools==4.56.0
functions-framework==3.8.2
google-api-core==2.24.1
google-api-python-client==1.7.11
google-auth==2.38.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==1.2.1
google-cloud-bigquery==3.29.0
google-cloud-bigquery-storage==2.28.0
google-cloud-core==2.4.1
google-cloud-firestore==2.20.0
google-cloud-secret-manager==2.7.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
googleauthentication==0.0.18
greenlet==3.1.1
grpc-google-iam-v1==0.12.7
grpcio==1.70.0
grpcio-status==1.62.3
h11==0.14.0
holidays==0.67
httplib2==0.22.0
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
jsonify==0.5
jwt==1.3.1
kiwisolver==1.4.8
libcst==1.6.0
MarkupSafe==3.0.2
matplotlib==3.10.0
Naked==0.1.32
numpy==2.2.2
oauthlib==3.2.2
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
passlib==1.7.4
patsy==1.0.1
pillow==11.1.0
prophet==1.1.6
proto-plus==1.26.0
protobuf==4.25.6
pyarrow==19.0.0
pyasn1==0.4.8
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.28.1
shellescape==3.8.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.38
stanio==0.5.1
starlette==0.46.1
statsmodels==0.14.4
threadpoolctl==3.5.0
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==3.0.1
urllib3==2.3.0
uvicorn==0.34.0
watchdog==6.0.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
